fixed all warnings

+++ oscillator.cpp
@@ -10,9 +10,9 @@ Oscillator::Oscillator(Wavedex w, std::unique_ptr<Frequency> f)
   : wavedex(w), frequency(std::move(f)) {};
 
 int16_t Oscillator::next(int32_t tick, int32_t phi) const {
-  uint32_t freq = frequency->get_frequency();
+  int64_t freq = frequency->get_frequency();
   // convert phi from simple_min-sample_max to -pi-pi (kinda)
-  int64_t phi_tmp = phi * (int64_t)freq;
+  int64_t phi_tmp = phi * freq;
   // fudge allows more variation (phi limited to sample_max)
   // but may need to worry about gain sensitivity
   int32_t phi_frac = phi_tmp >> (sample_bits - 1 - phi_fudge_bits);
